move NativeFindStorageItemHelper to backend
