+++ src/Files.Uwp/Filesystem/StorageItems/VirtualStorageItem.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 using Windows.Foundation;
 using Windows.Storage;
 using Windows.Storage.FileProperties;
-using static Files.Uwp.Helpers.NativeFindStorageItemHelper;
+using static Files.Backend.Helpers.NativeFindStorageItemHelper;
 
 namespace Files.Uwp.Filesystem.StorageItems
 {
