+++ src/Files.Uwp/Helpers/NavigationHelpers.cs
@@ -1,9 +1,10 @@
 ﻿using CommunityToolkit.Mvvm.DependencyInjection;
+using Files.Backend.Helpers;
 using Files.Backend.Services.Settings;
-using Files.Uwp.Filesystem;
-using Files.Uwp.Filesystem.StorageItems;
 using Files.Shared;
 using Files.Shared.Enums;
+using Files.Uwp.Filesystem;
+using Files.Uwp.Filesystem.StorageItems;
 using Files.Uwp.ViewModels;
 using Files.Uwp.Views;
 using Microsoft.Toolkit.Uwp;
